repo reorg: move echo filer and integration test
