+++ source/extensions/BUILD
@@ -1,0 +1,1 @@
+licenses(["notice"])  # Apache 2
