--- source/server/config/network/echo.cc
@@ -1,46 +1,0 @@
-#include <string>
-
-#include "envoy/registry/registry.h"
-#include "envoy/server/filter_config.h"
-
-#include "common/config/well_known_names.h"
-#include "common/filter/echo.h"
-
-namespace Envoy {
-namespace Server {
-namespace Configuration {
-
-/**
- * Config registration for the echo filter. @see NamedNetworkFilterConfigFactory.
- */
-class EchoConfigFactory : public NamedNetworkFilterConfigFactory {
-public:
-  // NamedNetworkFilterConfigFactory
-  NetworkFilterFactoryCb createFilterFactory(const Json::Object&, FactoryContext&) override {
-    return [](Network::FilterManager& filter_manager) -> void {
-      filter_manager.addReadFilter(Network::ReadFilterSharedPtr{new Filter::Echo()});
-    };
-  }
-
-  NetworkFilterFactoryCb createFilterFactoryFromProto(const Protobuf::Message&,
-                                                      FactoryContext&) override {
-    return [](Network::FilterManager& filter_manager) -> void {
-      filter_manager.addReadFilter(Network::ReadFilterSharedPtr{new Filter::Echo()});
-    };
-  }
-
-  ProtobufTypes::MessagePtr createEmptyConfigProto() override {
-    return ProtobufTypes::MessagePtr{new Envoy::ProtobufWkt::Empty()};
-  }
-
-  std::string name() override { return Config::NetworkFilterNames::get().ECHO; }
-};
-
-/**
- * Static registration for the echo filter. @see RegisterFactory.
- */
-static Registry::RegisterFactory<EchoConfigFactory, NamedNetworkFilterConfigFactory> registered_;
-
-} // namespace Configuration
-} // namespace Server
-} // namespace Envoy
